# back/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /store:
    post:
      description: Cria uma nova loja para o produtor autenticado
      operationId: StoreController_create
      parameters: []
      requestBody:
        required: true
        description: Dados da loja a ser criada
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreDto'
            examples:
              storeExample:
                summary: Exemplo de loja
                description: Dados básicos para criação de loja
                value:
                  name: Fazenda Orgânica
                  description: Produtos orgânicos cultivados com amor
                  producerProfileId: 1
              storeWithStockExample:
                summary: Loja com estoque inicial
                description: Criar loja com configuração inicial de estoque
                value:
                  name: Horta Feliz
                  description: Produtos frescos direto da horta
                  producerProfileId: 1
                  stock:
                    name: Estoque principal
                    description: Estoque inicial da Horta Feliz
      responses:
        '201':
          description: Loja criada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para criar lojas
      security:
        - JWT-auth: []
      summary: Criar uma nova loja
      tags: &ref_3
        - Store
    get:
      description: Retorna uma lista com todas as lojas cadastradas
      operationId: StoreController_findAll
      parameters: []
      responses:
        '200':
          description: Lista de lojas retornada com sucesso
      summary: Listar todas as lojas
      tags: *ref_3
  /store/{id}:
    get:
      description: Retorna os dados de uma loja específica
      operationId: StoreController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      responses:
        '200':
          description: Loja encontrada com sucesso
        '404':
          description: Loja não encontrada
      summary: Obter loja por ID
      tags: *ref_3
    patch:
      description: Atualiza os dados de uma loja específica
      operationId: StoreController_update
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      requestBody:
        required: true
        description: Dados para atualização da loja
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreDto'
            examples:
              updateInfoExample:
                summary: Atualizar informações básicas
                description: Atualizar nome e descrição da loja
                value:
                  name: Novo Nome da Loja
                  description: Nova descrição da loja atualizada
              updateRatingExample:
                summary: Atualizar avaliação
                description: Atualizar apenas a avaliação da loja
                value:
                  rating: 4.5
      responses:
        '200':
          description: Loja atualizada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para atualizar esta loja
        '404':
          description: Loja não encontrada
      security:
        - JWT-auth: []
      summary: Atualizar loja
      tags: *ref_3
    delete:
      description: Remove uma loja específica
      operationId: StoreController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      responses:
        '200':
          description: Loja removida com sucesso
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para remover esta loja
        '404':
          description: Loja não encontrada
      security:
        - JWT-auth: []
      summary: Remover loja
      tags: *ref_3
  /store/{storeId}/products:
    get:
      description: Retorna todos os produtos disponíveis em uma loja específica
      operationId: StoreController_findProductsByStoreId
      parameters:
        - name: storeId
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
        '404':
          description: Loja não encontrada
      summary: Listar produtos de uma loja
      tags: *ref_3
  /store/{id}/stock:
    get:
      description: Retorna informações do estoque de uma loja específica
      operationId: StoreController_getStock
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      responses:
        '200':
          description: Estoque encontrado com sucesso
        '404':
          description: Loja ou estoque não encontrado
      summary: Obter estoque da loja
      tags: *ref_3
...
info:
  title: Ze da Horta API
  description: API documentation for the NestJS application
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Insira seu token JWT
      in: header
  schemas:
...
    CreateStoreDto:
      type: object
      properties: {}
    UpdateStoreDto:
      type: object
      properties: {}
...